Qute type-safe messages - default bundle name strategy #2
- follows up on https://github.com/quarkusio/quarkus/pull/31299
