+++ extensions/qute/deployment/src/test/java/io/quarkus/qute/deployment/i18n/MessageBundleDefaultedNameTest.java
@@ -16,17 +16,17 @@ public class MessageBundleDefaultedNameTest {
     @RegisterExtension
     static final QuarkusUnitTest config = new QuarkusUnitTest()
             .withApplicationRoot((jar) -> jar
-                    .addClasses(Views.class)
+                    .addClasses(Controller.class)
                     .addAsResource(new StringAsset(
-                            "{msg_Views_Index:hello(name)}"),
-                            "templates/Index/index.html")
-                    .addAsResource(new StringAsset("hello=Ahoj {name}!"), "messages/msg_Views_Index_cs.properties"));
+                            "{Controller_index:hello(name)}"),
+                            "templates/Controller/index.html")
+                    .addAsResource(new StringAsset("hello=Ahoj {name}!"), "messages/Controller_index_cs.properties"));
 
     @Test
     public void testBundle() {
         assertEquals("Hello world!",
-                Views.Index.Templates.index("world").render());
-        assertEquals("Ahoj svete!", Views.Index.Templates.index("svete")
+                Controller.Templates.index("world").render());
+        assertEquals("Ahoj svete!", Controller.Templates.index("svete")
                 .setAttribute(MessageBundles.ATTRIBUTE_LOCALE, Locale.forLanguageTag("cs")).render());
     }
